Beginning rewrite of MLP
I've had an issue with the model not training. And I've found that some of the things I've been using are not optimal for this, such as sigmoid instead of ReLu or mean squared error loss on a multi class classification problem instead of cross-entropy. Im also dissatisfied with the way I've organized the MLP so Im doing a rewrite, taking whatever I can from MLP file but using different functions and organization